Refactor future-fetch
